requirement.txt
django
whitenoise
gunicorn
Pillow
django-stdimage